requirements.txt
Flask==3.0.3
PyMySQL==1.1.1
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2          # ✅ pinned to avoid passlib error
pyotp==2.9.0
python-dotenv==1.0.1
flask-cors==4.0.1
uvicorn==0.30.1        # ✅ for ASGI server
asgiref==3.8.1         # ✅ needed for WSGI → ASGI adapter
